backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Session==0.5.0
python-dotenv==1.0.0
Werkzeug==3.0.1
mysql-connector-python==8.2.0
Pillow==10.1.0
